requirements.txt
requests
beautifulsoup4
google-api-python-client
google-auth-oauthlib
